+++ packages/admin/brevo-admin/CHANGELOG.md
@@ -1,6 +1,5 @@
 # @comet/brevo-admin
 
-<<<<<<< HEAD
 ## 9.0.0-beta.0
 
 ### Minor Changes
@@ -21,7 +20,7 @@
     - @comet/admin-icons@9.0.0-beta.0
     - @comet/admin@9.0.0-beta.0
     - @comet/cms-admin@9.0.0-beta.0
-=======
+
 ## 8.20.0
 
 ### Patch Changes
@@ -57,7 +56,6 @@
     - @comet/admin@8.18.0
     - @comet/admin-date-time@8.18.0
     - @comet/admin-icons@8.18.0
->>>>>>> main
 
 ## 8.17.1