+++ packages/admin/brevo-admin/CHANGELOG.md
@@ -21,6 +21,42 @@
     - @comet/admin@9.0.0-beta.0
     - @comet/cms-admin@9.0.0-beta.0
 
+## 8.20.0
+
+### Patch Changes
+
+- Updated dependencies [ed00704]
+- Updated dependencies [412ed19]
+- Updated dependencies [caceff8]
+    - @comet/cms-admin@8.20.0
+    - @comet/admin@8.20.0
+    - @comet/admin-icons@8.20.0
+    - @comet/admin-date-time@8.20.0
+
+## 8.19.0
+
+### Patch Changes
+
+- Updated dependencies [fff2cc2]
+- Updated dependencies [fff2cc2]
+    - @comet/admin@8.19.0
+    - @comet/cms-admin@8.19.0
+    - @comet/admin-date-time@8.19.0
+    - @comet/admin-icons@8.19.0
+
+## 8.18.0
+
+### Patch Changes
+
+- Updated dependencies [64b70bc]
+- Updated dependencies [ed029fb]
+- Updated dependencies [0ce431c]
+- Updated dependencies [d344f53]
+    - @comet/cms-admin@8.18.0
+    - @comet/admin@8.18.0
+    - @comet/admin-date-time@8.18.0
+    - @comet/admin-icons@8.18.0
+
 ## 8.17.1
 
 ### Patch Changes